backend/requirements.txt
flask
flask-socketio
speechrecognition
torch
transformers
datasets
sentencepiece
openai-whisper
opencv-python
pydub
